Fail attempt for making robustness efficient.
1. Made horizon extremely conservative, i.e., actual horizon <= computed
   horizon.

   (TODO: check with Bardh and Mohammad if the formula makes sense)

2. Tried to make a table for computing robustness efficiently in an
   online manner using std::shared_ptr<const Expr>. This was a huge
   failure, with several pages of compiler errors.

   See default_monitor.cc for a little bit more detailed notes.

Minor changes
-------------

I am going to be using cppitertools which seems to be a good third-party
project with Python-like itertools. This will be useful in constructing
the Cartesian product of Var_id that have been pinned in the subformula
tree.

Extra Notes
-----------

The current implementation of RobustnessOp cannot be parallelized as the
RobustnessOp struct holds the context (Var maps) for the subformulas,
which will be updated on what we see in each subformula.

In my opinion, the tabular, DP-based formulation also cannot be
parallelized without some thread wrangling. One thing that comes to my
minds is using a Task Graph, where each node in the Graph holds its
required robustness buffer -- although I am not sure how we can update
teh buffer efficiently.