fix shifting eval context to the past
Previously, there were segfaults due to the sentinel being wrong in the
call to `ranges::prev` and because I was using a value past the sentinel
as the current frame.

The fix was to first use the proper `begin` sentinel, and then (since
`by` is non-zero) use the sentinel itself as the current frame (as that
is the "1 past the history" frame).

The second thing may cause an issue when there is an empty history and
an operator tries to reference the current frame, but in general, the
shift should not occur when there is an empty history because the
temporal operator should have a guard that returns vacuously.